// src/types.ts
export enum Role {
  ADMIN = 'ADMIN',
  STAFF = 'STAFF',
  MANAGER = 'MANAGER',
  SUPER_ADMIN = 'SUPER_ADMIN', // New Role for System Owner
}

export enum RoomStatus {
  AVAILABLE = 'AVAILABLE', // Green
  OCCUPIED = 'OCCUPIED',   // Red
  PAYMENT = 'PAYMENT',     // Blue (New status)
  CLEANING = 'CLEANING',   // Yellow
  ERROR = 'ERROR',         // Black
}
// ...
export interface Room {
  id: string;
  storeId: string; // Linked to Store
  name: string;
  status: RoomStatus;
  type: 'VIP' | 'NORMAL';
  hourlyRate: number;
  currentOrderId?: string;
  checkInTime?: number; // Timestamp
  imageUrl?: string; // New field for room image
}